Conversion of BinarySerialier to Profile37 PCL.
In addition to converting the project, several code changes were
necessary to remove dependencies on .NET only APIs. This included:
Stream does not provide Close method. Instead derived classes should
override the Dispose method.
The MemberInfo class in the Reflection API does not have a MemberType
property. Instead, the code attempts to cast the object to a
PropertyInfo or FieldInfo and perform the actions based on the success
of the cast.
PCLs do not support ComVisibleAttribute or GuidAttribute.
PCL versions of the Convert and Encoding classes do not have overrides
previous in use.